# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Persome API",
    "description": "Local-first screen-context memory and personal-model REST API.",
    "version": "0.2.2"
  },
  "paths": {
...
    "/captures/ingest": {
      "post": {
        "tags": [
          "capture"
        ],
        "summary": "Ingest Capture",
        "description": "Ingest one frame from a trusted local producer.\n\nThis replay/embedding path shares enrichment, deduplication, and persistence\nwith the daemon's native macOS AX watcher.",
        "operationId": "ingest_capture_captures_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureIngestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/model/node": {
      "get": {
        "tags": [
          "model"
        ],
        "summary": "Model Node",
        "description": "Return the symbolic receipts behind one graph node.\n\nThis is the click-through from the visual graph to the symbolic layer. Lazy per-node fetch\nso the graph payload stays lean. Zero-LLM, read-only, fail-open:\n\n- ``event:<source-kind>:<source-id>`` → the canonical Activity source;\n- any other identity → the latest ACTIVE evo_nodes of its typed entity file\n  entity file (newest first, bounded, truncated) — the consolidation\n  trail the point was distilled from.",
        "operationId": "model_node_model_node_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Model Node Model Node Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the request succeeded",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload"
          }
        },
        "type": "object",
        "title": "ApiResponse",
        "description": "Common response envelope for the local runtime API."
      },
      "CaptureIngestBody": {
        "properties": {
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "ISO 8601 capture time; generated by the server when omitted"
          },
          "trigger": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger",
            "description": "Trigger metadata: {event_type, bundle_id, window_title, details?}"
          },
          "window_meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Window Meta",
            "description": "{app_name, title, bundle_id}"
          },
          "ax_tree": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ax Tree",
            "description": "Raw AX tree: {apps -> windows -> elements}; omission means AX is unavailable"
          },
          "ax_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ax Metadata",
            "description": "AX capture metadata"
          },
          "screenshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screenshot",
            "description": "{image_base64, mime_type, width, height}; plaintext in transit and encrypted at rest according to server configuration"
          },
          "ocr_jpeg_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Jpeg B64",
            "description": "Base64 focused-window JPEG for local OCR when AX text is insufficient"
          },
          "ocr_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Tier",
            "description": "Per-capture OCR tier override; uses configuration when omitted"
          }
        },
        "type": "object",
        "title": "CaptureIngestBody",
        "description": "One capture frame pushed by a trusted Swift producer.\n\nThe Swift process owns Accessibility and Screen Recording permissions and\nsupplies the AX tree plus an optional focused-window screenshot. The daemon\nperforms enrichment and persistence. Fields align with the native\n``_build_capture`` output and degrade safely when optional data is absent."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}