# src/persome/api/models.py
"""Pydantic models for the Persome HTTP REST API."""

from __future__ import annotations

from typing import Any

from pydantic import BaseModel, Field

# ─── Generic envelope ──────────────────────────────────────────────────────


class ApiResponse(BaseModel):
    """Common response envelope for the local runtime API."""

    success: bool = Field(default=True, description="Whether the request succeeded")
    data: dict | list | str | None = Field(default=None, description="Response payload")


class ModelPing(BaseModel):
    stage: str = Field(description="LLM stage name, such as timeline, reducer, or classifier")
    model: str = Field(description="Configured model name")
    ok: bool = Field(description="Whether the stage model is reachable")
    latency_ms: int | None = Field(default=None, description="Ping latency in milliseconds")
    error: str | None = Field(default=None, description="Error returned when the ping fails")


class CaptureIngestBody(BaseModel):
    """One capture frame pushed by a trusted Swift producer.

    The Swift process owns Accessibility and Screen Recording permissions and
    supplies the AX tree plus an optional focused-window screenshot. The daemon
    performs enrichment and persistence. Fields align with the native
    ``_build_capture`` output and degrade safely when optional data is absent.
    """

    timestamp: str | None = Field(
        default=None, description="ISO 8601 capture time; generated by the server when omitted"
    )
    trigger: dict[str, Any] | None = Field(
        default=None,
        description="Trigger metadata: {event_type, bundle_id, window_title, details?}",
    )
    window_meta: dict[str, Any] = Field(
        default_factory=dict, description="{app_name, title, bundle_id}"
    )
    ax_tree: dict[str, Any] | None = Field(
        default=None,
        description="Raw AX tree: {apps -> windows -> elements}; omission means AX is unavailable",
    )
    ax_metadata: dict[str, Any] | None = Field(default=None, description="AX capture metadata")
    screenshot: dict[str, Any] | None = Field(
        default=None,
        description=(
            "{image_base64, mime_type, width, height}; plaintext in transit and encrypted "
            "at rest according to server configuration"
        ),
    )
    ocr_jpeg_b64: str | None = Field(
        default=None,
        description="Base64 focused-window JPEG for local OCR when AX text is insufficient",
    )
    ocr_tier: str | None = Field(
        default=None, description="Per-capture OCR tier override; uses configuration when omitted"
    )